Move the following functions from base.py to auxiliary.py
- get_run_deferred_hooks_and_restart_pre_upgrade_step
- get_run_deferred_hooks_and_restart_post_upgrade_step
